+++ lib/timers.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const Timer = process.binding('timer_wrap').Timer;
-const L = require('_linklist');
+const L = require('internal/linkedlist');
 const assert = require('assert').ok;
 const util = require('util');
 const debug = util.debuglog('timer');
