+++ node.gyp
@@ -17,7 +17,6 @@
       'src/node.js',
       'lib/_debug_agent.js',
       'lib/_debugger.js',
-      'lib/_linklist.js',
       'lib/assert.js',
       'lib/buffer.js',
       'lib/child_process.js',
@@ -39,6 +38,7 @@
       'lib/_http_outgoing.js',
       'lib/_http_server.js',
       'lib/https.js',
+      'lib/_linklist.js',
       'lib/module.js',
       'lib/net.js',
       'lib/os.js',
@@ -70,9 +70,10 @@
       'lib/zlib.js',
       'lib/internal/child_process.js',
       'lib/internal/freelist.js',
+      'lib/internal/linkedlist.js',
       'lib/internal/module.js',
-      'lib/internal/socket_list.js',
       'lib/internal/repl.js',
+      'lib/internal/socket_list.js',
       'lib/internal/util.js',
       'lib/internal/streams/lazy_transform.js',
     ],
